fix: refactored notifications component
